@@ -4,24 +4,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^1.6.0",
-    "bootstrap": "^5.1.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.0-rc.0",
-    "react-dom": "^17.0.2",
+    "@fortawesome/fontawesome-svg-core": "^6.6.0",
+    "@reduxjs/toolkit": "^2.2.8",
+    "@testing-library/jest-dom": "6.5.0",
+    "@testing-library/react": "16.0.1",
+    "@testing-library/user-event": "14.5.2",
+    "axios": "1.7.7",
+    "bootstrap": "5.3.3",
+    "react": "18.3.1",
+    "react-bootstrap": "2.10.5",
+    "react-dom": "18.3.1",
     "react-helmet": "^6.1.0",
     "react-paypal-button-v2": "^2.6.3",
-    "react-redux": "^7.2.6",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.3.0",
+    "react-redux": "9.1.2",
+    "react-router-bootstrap": "0.26.3",
+    "react-router-dom": "6.27.0",
     "react-scripts": "5.0.1",
-    "redux": "^4.1.2",
+    "redux": "5.0.1",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.4.0",
-    "web-vitals": "^1.1.2"
+    "redux-thunk": "^3.1.0",
+    "web-vitals": "4.2.3"
   },
   "scripts": {
     "start": "react-scripts start",
