@@ -42,18 +42,24 @@ const getProductById = asyncHandle(async(req,res)=>{
 //@dec       Delete product
 //@route     DELETE /api/products/:id
 //@access    Private /admin
-const deleteProduct = asyncHandle(async(req,res)=>{
-
-    const product =await Product.findById(req.params.id)
-
+const deleteProduct = asyncHandle(async (req, res) => {
+  
+  try {
+    const product = await Product.findByIdAndDelete(req.params.id);
+    
     if (product) {
-        await product.remove()
-        res.send({message: "Product deleted successfully"})
-    }else{
-        res.status(404)
-        throw new Error("Product Not Found")
+      return res.json({ message: 'Product removed' });
+    } else {
+      return res.status(404).json({ message: 'Product not found' });
     }
-})
+  } catch (error) {
+    console.error(`Error deleting product: ${error.message}`);
+    return res.status(500).json({ message: 'Internal server error' });
+  }
+});
+
+
+
 
 //@dec       Create a product
 //@route     POST /api/products
