@@ -1,20 +1,24 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Form, Button } from 'react-bootstrap';
 import { useDispatch, useSelector } from 'react-redux';
 import FormContainer from '../Components/formContainer';
 import CheckoutSteps from '../Components/CheckoutSteps';
 import { savePaymentMethod } from '../actions/cartActions';
+import { useNavigate } from 'react-router-dom'; // Import useNavigate
 
-const PaymentScreen = ({ history }) => {
+const PaymentScreen = () => {
     const cart = useSelector((state) => state.cart);
     const { shippingAddress } = cart;
 
-    if (!shippingAddress.address) {
-        history.push('/shipping');
-    }
+    const navigate = useNavigate(); // Initialize navigate function
+
+    useEffect(() => {
+        if (!shippingAddress.address) {
+            navigate('/shipping'); // Redirect if no shipping address
+        }
+    }, [shippingAddress, navigate]);
 
     const [paymentMethod, setPaymentMethod] = useState('PayPal');
-
     const dispatch = useDispatch();
 
     const handleSelect = (method) => {
@@ -24,7 +28,7 @@ const PaymentScreen = ({ history }) => {
     const submitHandler = (e) => {
         e.preventDefault();
         dispatch(savePaymentMethod(paymentMethod));
-        history.push('/placeorder');
+        navigate('/placeorder'); // Use navigate for redirection
     };
 
     return (
@@ -38,54 +42,25 @@ const PaymentScreen = ({ history }) => {
                         <h1 style={{ textAlign: 'center', fontSize: '24px', color: '#333', marginBottom: '20px' }}>Payment Method</h1>
                         <Form onSubmit={submitHandler}>
                             <Form.Group>
-                                <div
-                                    style={{ background: paymentMethod === 'PayPal' ? '#007bff' : '#fff', padding: '20px', borderRadius: '10px', boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)', marginBottom: '20px', cursor: 'pointer' }}
-                                    onClick={() => handleSelect('PayPal')}
-                                    className="payment-tile"
-                                >
-                                    <Form.Check
-                                        type='radio'
-                                        label='PayPal or Credit Card'
-                                        id='PayPal'
-                                        name='paymentMethod'
-                                        value='PayPal'
-                                        checked={paymentMethod === 'PayPal'}
-                                        onChange={() => handleSelect('PayPal')}
-                                        style={{ fontSize: '16px', color: paymentMethod === 'PayPal' ? '#fff' : '#333' }}
-                                    />
-                                </div>
-                                <div
-                                    style={{ background: paymentMethod === 'UPI' ? '#007bff' : '#fff', padding: '20px', borderRadius: '10px', boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)', marginBottom: '20px', cursor: 'pointer' }}
-                                    onClick={() => handleSelect('UPI')}
-                                    className="payment-tile"
-                                >
-                                    <Form.Check
-                                        type='radio'
-                                        label='UPI (Unified Payments Interface)'
-                                        id='UPI'
-                                        name='paymentMethod'
-                                        value='UPI'
-                                        checked={paymentMethod === 'UPI'}
-                                        onChange={() => handleSelect('UPI')}
-                                        style={{ fontSize: '16px', color: paymentMethod === 'UPI' ? '#fff' : '#333' }}
-                                    />
-                                </div>
-                                <div
-                                    style={{ background: paymentMethod === 'Cash' ? '#007bff' : '#fff', padding: '20px', borderRadius: '10px', boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)', marginBottom: '20px', cursor: 'pointer' }}
-                                    onClick={() => handleSelect('Cash')}
-                                    className="payment-tile"
-                                >
-                                    <Form.Check
-                                        type='radio'
-                                        label='Cash on Delivery'
-                                        id='Cash'
-                                        name='paymentMethod'
-                                        value='Cash'
-                                        checked={paymentMethod === 'Cash'}
-                                        onChange={() => handleSelect('Cash')}
-                                        style={{ fontSize: '16px', color: paymentMethod === 'Cash' ? '#fff' : '#333' }}
-                                    />
-                                </div>
+                                {['PayPal', 'UPI', 'Cash'].map((method) => (
+                                    <div
+                                        key={method}
+                                        style={{ background: paymentMethod === method ? '#007bff' : '#fff', padding: '20px', borderRadius: '10px', boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)', marginBottom: '20px', cursor: 'pointer' }}
+                                        onClick={() => handleSelect(method)}
+                                        className="payment-tile"
+                                    >
+                                        <Form.Check
+                                            type='radio'
+                                            label={method === 'PayPal' ? 'PayPal or Credit Card' : method === 'UPI' ? 'UPI (Unified Payments Interface)' : 'Cash on Delivery'}
+                                            id={method}
+                                            name='paymentMethod'
+                                            value={method}
+                                            checked={paymentMethod === method}
+                                            onChange={() => handleSelect(method)}
+                                            style={{ fontSize: '16px', color: paymentMethod === method ? '#fff' : '#333' }}
+                                        />
+                                    </div>
+                                ))}
                             </Form.Group>
 
                             <Button type='submit' variant='primary' className="w-100" style={{ fontSize: '16px', marginTop: '20px', background: '#007bff', border: 'none', borderRadius: '5px', padding: '10px 20px' }}>
